# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Industrial Contract & Supplier Lifecycle Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User deleted
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List suppliers
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierRequest"
      responses:
        "201":
          description: Supplier created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get supplier by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    put:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSupplierRequest"
      responses:
        "200":
          description: Supplier updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents (filtered by vendor for suppliers)
      parameters:
        - name: supplierId
          in: query
          schema:
            type: integer
        - name: contractId
          in: query
          schema:
            type: integer
        - name: documentType
          in: query
          schema:
            type: string
            enum: [contract, technical_drawing, invoice, compliance_cert, sop, other]
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
  /documents/upload:
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                supplierId:
                  type: integer
                contractId:
                  type: integer
                documentType:
                  type: string
                  enum: [contract, technical_drawing, invoice, compliance_cert, sop, other]
                description:
                  type: string
              required: [file, supplierId, documentType]
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Document deleted
  /documents/{id}/download:
    get:
      operationId: downloadDocument
      tags: [documents]
      summary: Download a document (logs audit trail)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, supplier, client]
        supplierId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, role, createdAt]
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, supplier, client]
        supplierId:
          type: integer
      required: [email, name, password, role]
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, supplier, client]
        supplierId:
          type: integer
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive, suspended]
        complianceScore:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, code, contactEmail, status, createdAt]
    CreateSupplierRequest:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [active, inactive, suspended]
      required: [name, code, contactEmail]
    UpdateSupplierRequest:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [active, inactive, suspended]
        complianceScore:
          type: number
...
    Document:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        originalName:
          type: string
        supplierId:
          type: integer
        contractId:
          type: integer
          nullable: true
        documentType:
          type: string
          enum: [contract, technical_drawing, invoice, compliance_cert, sop, other]
        description:
          type: string
          nullable: true
        fileSize:
          type: integer
        mimeType:
          type: string
        uploadedBy:
          type: integer
        uploaderName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, filename, originalName, supplierId, documentType, fileSize, mimeType, uploadedBy, createdAt]
...