# lib/api-zod/src/generated/types/createSupplierRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Industrial Contract & Supplier Lifecycle Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSupplierRequestStatus } from "./createSupplierRequestStatus";

export interface CreateSupplierRequest {
  name: string;
  code: string;
  contactEmail: string;
  contactPhone?: string;
  address?: string;
  status?: CreateSupplierRequestStatus;
}