# lib/api-zod/src/generated/types/createUserRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Industrial Contract & Supplier Lifecycle Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateUserRequestRole } from "./createUserRequestRole";

export interface CreateUserRequest {
  email: string;
  name: string;
  password: string;
  role: CreateUserRequestRole;
  supplierId?: number;
}